refactor: wallet packages
