+++ wallet-service/src/main/java/com/ada/tech/wallet/WalletServiceApplication.java
@@ -1,0 +1,13 @@
+package com.ada.tech.wallet;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class WalletServiceApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(WalletServiceApplication.class, args);
+    }
+}
+
