+++ wallet-service/src/main/java/com/ada/tech/wallet/config/OpenAPIConfiguration.java
@@ -1,0 +1,30 @@
+package com.ada.tech.wallet.config;
+
+import io.swagger.v3.oas.models.OpenAPI;
+import io.swagger.v3.oas.models.info.Info;
+import org.springdoc.core.models.GroupedOpenApi;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class OpenAPIConfiguration {
+
+    @Bean
+    public OpenAPI walletOpenAPI() {
+        return new OpenAPI()
+                .info(new Info()
+                        .title("Wallet Service API")
+                        .version("1.0.0")
+                        .description("API for managing digital wallets.")
+                );
+    }
+
+    @Bean
+    public GroupedOpenApi walletApi() {
+        return GroupedOpenApi.builder()
+                .group("wallet-service")
+                .pathsToMatch("/api/**")
+                .build();
+    }
+
+}
