+++ wallet-service/src/main/java/com/ada/tech/wallet/dto/CreateWalletRequest.java
@@ -1,0 +1,18 @@
+package com.ada.tech.wallet.dto;
+
+import jakarta.validation.constraints.NotNull;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class CreateWalletRequest {
+    
+    @NotNull(message = "User ID is required")
+    private Long userId;
+}
+
