+++ wallet-service/src/main/java/com/ada/tech/wallet/dto/WithdrawalRequest.java
@@ -1,0 +1,24 @@
+package com.ada.tech.wallet.dto;
+
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Positive;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.math.BigDecimal;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class WithdrawalRequest {
+    
+    @NotNull(message = "Amount is required")
+    @Positive(message = "Amount must be positive")
+    private BigDecimal amount;
+    
+    private String description;
+}
+
