+++ wallet-service/src/main/java/com/ada/tech/wallet/exception/InsufficientFundsException.java
@@ -1,0 +1,13 @@
+package com.ada.tech.wallet.exception;
+
+public class InsufficientFundsException extends RuntimeException {
+    
+    public InsufficientFundsException(String message) {
+        super(message);
+    }
+    
+    public InsufficientFundsException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
+
