+++ wallet-service/src/main/java/com/ada/tech/wallet/exception/TransactionFailedException.java
@@ -1,0 +1,13 @@
+package com.ada.tech.wallet.exception;
+
+public class TransactionFailedException extends RuntimeException {
+    
+    public TransactionFailedException(String message) {
+        super(message);
+    }
+    
+    public TransactionFailedException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
+
