+++ wallet-service/src/main/java/com/ada/tech/wallet/exception/WalletAlreadyExistsException.java
@@ -1,0 +1,13 @@
+package com.ada.tech.wallet.exception;
+
+public class WalletAlreadyExistsException extends RuntimeException {
+    
+    public WalletAlreadyExistsException(String message) {
+        super(message);
+    }
+    
+    public WalletAlreadyExistsException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
+
