+++ wallet-service/src/main/java/com/ada/tech/wallet/exception/WalletNotFoundException.java
@@ -1,0 +1,13 @@
+package com.ada.tech.wallet.exception;
+
+public class WalletNotFoundException extends RuntimeException {
+    
+    public WalletNotFoundException(String message) {
+        super(message);
+    }
+    
+    public WalletNotFoundException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
+
