+++ wallet-service/src/main/java/com/ada/tech/wallet/model/TransactionStatus.java
@@ -1,0 +1,8 @@
+package com.ada.tech.wallet.model;
+
+public enum TransactionStatus {
+    PENDING,
+    COMPLETED,
+    FAILED
+}
+
