+++ wallet-service/src/main/java/com/ada/tech/wallet/model/TransactionType.java
@@ -1,0 +1,9 @@
+package com.ada.tech.wallet.model;
+
+public enum TransactionType {
+    DEPOSIT,
+    WITHDRAWAL,
+    TRANSFER_IN,
+    TRANSFER_OUT
+}
+
