+++ wallet-service/src/main/java/com/ada/tech/wallet/repository/WalletRepository.java
@@ -1,0 +1,14 @@
+package com.ada.tech.wallet.repository;
+
+import com.ada.tech.wallet.model.Wallet;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.UUID;
+
+@Repository
+public interface WalletRepository extends JpaRepository<Wallet, UUID> {
+    
+    boolean existsByUserId(Long userId);
+}
+
