--- wallet-service/src/main/java/com/wallet/service/WalletServiceApplication.java
@@ -1,13 +1,0 @@
-package com.wallet.service;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class WalletServiceApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(WalletServiceApplication.class, args);
-    }
-}
-
