--- wallet-service/src/main/java/com/wallet/service/config/OpenAPIConfiguration.java
@@ -1,30 +1,0 @@
-package com.wallet.service.config;
-
-import io.swagger.v3.oas.models.OpenAPI;
-import io.swagger.v3.oas.models.info.Info;
-import org.springdoc.core.models.GroupedOpenApi;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class OpenAPIConfiguration {
-
-    @Bean
-    public OpenAPI walletOpenAPI() {
-        return new OpenAPI()
-                .info(new Info()
-                        .title("Wallet Service API")
-                        .version("1.0.0")
-                        .description("API for managing digital wallets.")
-                );
-    }
-
-    @Bean
-    public GroupedOpenApi walletApi() {
-        return GroupedOpenApi.builder()
-                .group("wallet-service")
-                .pathsToMatch("/api/**")
-                .build();
-    }
-
-}
