--- wallet-service/src/main/java/com/wallet/service/dto/BalanceResponse.java
@@ -1,24 +1,0 @@
-package com.wallet.service.dto;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.math.BigDecimal;
-import java.time.Instant;
-import java.util.UUID;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-@JsonInclude(JsonInclude.Include.NON_NULL)
-public class BalanceResponse {
-    
-    private UUID walletId;
-    private BigDecimal balance;
-    private Instant timestamp;
-}
-
