--- wallet-service/src/main/java/com/wallet/service/dto/CreateWalletRequest.java
@@ -1,18 +1,0 @@
-package com.wallet.service.dto;
-
-import jakarta.validation.constraints.NotNull;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class CreateWalletRequest {
-    
-    @NotNull(message = "User ID is required")
-    private Long userId;
-}
-
