--- wallet-service/src/main/java/com/wallet/service/dto/TransferRequest.java
@@ -1,31 +1,0 @@
-package com.wallet.service.dto;
-
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Positive;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.math.BigDecimal;
-import java.util.UUID;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class TransferRequest {
-    
-    @NotNull(message = "Source wallet ID is required")
-    private UUID sourceWalletId;
-    
-    @NotNull(message = "Destination wallet ID is required")
-    private UUID destinationWalletId;
-    
-    @NotNull(message = "Amount is required")
-    @Positive(message = "Amount must be positive")
-    private BigDecimal amount;
-    
-    private String description;
-}
-
