--- wallet-service/src/main/java/com/wallet/service/exception/GlobalExceptionHandler.java
@@ -1,98 +1,0 @@
-package com.wallet.service.exception;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.validation.FieldError;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-
-import java.time.Instant;
-import java.util.HashMap;
-import java.util.Map;
-
-@ControllerAdvice
-@Slf4j
-public class GlobalExceptionHandler {
-
-    @ExceptionHandler(WalletNotFoundException.class)
-    public ResponseEntity<ErrorResponse> handleWalletNotFoundException(WalletNotFoundException ex) {
-        log.error("Wallet not found", ex);
-        return buildErrorResponse(HttpStatus.NOT_FOUND, ex.getMessage());
-    }
-
-    @ExceptionHandler(WalletAlreadyExistsException.class)
-    public ResponseEntity<ErrorResponse> handleWalletAlreadyExistsException(WalletAlreadyExistsException ex) {
-        log.error("Wallet already exists", ex);
-        return buildErrorResponse(HttpStatus.CONFLICT, ex.getMessage());
-    }
-
-    @ExceptionHandler(InsufficientFundsException.class)
-    public ResponseEntity<ErrorResponse> handleInsufficientFundsException(InsufficientFundsException ex) {
-        log.error("Insufficient funds", ex);
-        return buildErrorResponse(HttpStatus.BAD_REQUEST, ex.getMessage());
-    }
-
-    @ExceptionHandler(TransactionFailedException.class)
-    public ResponseEntity<ErrorResponse> handleTransactionFailedException(TransactionFailedException ex) {
-        log.error("Transaction failed", ex);
-        return buildErrorResponse(HttpStatus.INTERNAL_SERVER_ERROR, ex.getMessage());
-    }
-
-    @ExceptionHandler(MethodArgumentNotValidException.class)
-    public ResponseEntity<ValidationErrorResponse> handleValidationExceptions(MethodArgumentNotValidException ex) {
-        log.error("Validation error", ex);
-        
-        Map<String, String> errors = new HashMap<>();
-        ex.getBindingResult().getAllErrors().forEach(error -> {
-            String fieldName = ((FieldError) error).getField();
-            String errorMessage = error.getDefaultMessage();
-            errors.put(fieldName, errorMessage);
-        });
-        
-        ValidationErrorResponse response = new ValidationErrorResponse(
-                HttpStatus.BAD_REQUEST.value(),
-                "Validation error",
-                Instant.now(),
-                errors
-        );
-        
-        return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
-    }
-
-    @ExceptionHandler(Exception.class)
-    public ResponseEntity<ErrorResponse> handleGenericException(Exception ex) {
-        log.error("Unexpected error", ex);
-        return buildErrorResponse(HttpStatus.INTERNAL_SERVER_ERROR, "An unexpected error occurred");
-    }
-
-    private ResponseEntity<ErrorResponse> buildErrorResponse(HttpStatus status, String message) {
-        ErrorResponse response = new ErrorResponse(
-                status.value(),
-                message,
-                Instant.now()
-        );
-        return new ResponseEntity<>(response, status);
-    }
-
-    @Data
-    @AllArgsConstructor
-    public static class ErrorResponse {
-        private int status;
-        private String message;
-        private Instant timestamp;
-    }
-
-    @Data
-    @AllArgsConstructor
-    public static class ValidationErrorResponse {
-        private int status;
-        private String message;
-        private Instant timestamp;
-        private Map<String, String> errors;
-    }
-}
-
