--- wallet-service/src/main/java/com/wallet/service/exception/InsufficientFundsException.java
@@ -1,13 +1,0 @@
-package com.wallet.service.exception;
-
-public class InsufficientFundsException extends RuntimeException {
-    
-    public InsufficientFundsException(String message) {
-        super(message);
-    }
-    
-    public InsufficientFundsException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
-
