--- wallet-service/src/main/java/com/wallet/service/exception/TransactionFailedException.java
@@ -1,13 +1,0 @@
-package com.wallet.service.exception;
-
-public class TransactionFailedException extends RuntimeException {
-    
-    public TransactionFailedException(String message) {
-        super(message);
-    }
-    
-    public TransactionFailedException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
-
