--- wallet-service/src/main/java/com/wallet/service/exception/WalletAlreadyExistsException.java
@@ -1,13 +1,0 @@
-package com.wallet.service.exception;
-
-public class WalletAlreadyExistsException extends RuntimeException {
-    
-    public WalletAlreadyExistsException(String message) {
-        super(message);
-    }
-    
-    public WalletAlreadyExistsException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
-
