--- wallet-service/src/main/java/com/wallet/service/exception/WalletNotFoundException.java
@@ -1,13 +1,0 @@
-package com.wallet.service.exception;
-
-public class WalletNotFoundException extends RuntimeException {
-    
-    public WalletNotFoundException(String message) {
-        super(message);
-    }
-    
-    public WalletNotFoundException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
-
