--- wallet-service/src/main/java/com/wallet/service/model/Transaction.java
@@ -1,60 +1,0 @@
-package com.wallet.service.model;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.hibernate.annotations.CreationTimestamp;
-
-import java.math.BigDecimal;
-import java.time.Instant;
-import java.util.UUID;
-
-@Entity
-@Table(name = "transactions")
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class Transaction {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.UUID)
-    private UUID id;
-
-    @Column(name = "wallet_id", nullable = false)
-    private UUID walletId;
-
-    @Enumerated(EnumType.STRING)
-    @Column(nullable = false)
-    private TransactionType type;
-
-    @Column(nullable = false, precision = 19, scale = 4)
-    private BigDecimal amount;
-
-    @CreationTimestamp
-    @Column(name = "timestamp", nullable = false, updatable = false)
-    private Instant timestamp;
-
-    @Enumerated(EnumType.STRING)
-    @Column(nullable = false)
-    private TransactionStatus status;
-
-    @Column(name = "reference_id")
-    private UUID referenceId;
-
-    @Column
-    private String description;
-
-    @Column(name = "balance_after", nullable = false, precision = 19, scale = 4)
-    private BigDecimal balanceAfter;
-}
-
