--- wallet-service/src/main/java/com/wallet/service/model/TransactionStatus.java
@@ -1,8 +1,0 @@
-package com.wallet.service.model;
-
-public enum TransactionStatus {
-    PENDING,
-    COMPLETED,
-    FAILED
-}
-
