--- wallet-service/src/main/java/com/wallet/service/model/TransactionType.java
@@ -1,9 +1,0 @@
-package com.wallet.service.model;
-
-public enum TransactionType {
-    DEPOSIT,
-    WITHDRAWAL,
-    TRANSFER_IN,
-    TRANSFER_OUT
-}
-
