--- wallet-service/src/main/java/com/wallet/service/repository/WalletRepository.java
@@ -1,14 +1,0 @@
-package com.wallet.service.repository;
-
-import com.wallet.service.model.Wallet;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.UUID;
-
-@Repository
-public interface WalletRepository extends JpaRepository<Wallet, UUID> {
-    
-    boolean existsByUserId(Long userId);
-}
-
